:recycle: refactor: 카프카 및 클릭 클래스 계층에 맞게 이동
