+++ src/main/java/com/whereyouad/WhereYouAd/domains/click/domain/service/ClickEventProducer.java
@@ -1,4 +1,4 @@
-package com.whereyouad.WhereYouAd.infrastructure.client.click;
+package com.whereyouad.WhereYouAd.domains.click.domain.service;
 
 import com.whereyouad.WhereYouAd.domains.click.application.dto.ClickDto;
